Set up advanced logging system

+++ app/database/models.py
@@ -1,12 +1,12 @@
 from sqlalchemy.orm import declarative_base
-from sqlalchemy import Column, Integer, String, Boolean, Time
+from sqlalchemy import Column, Integer, String, Boolean, Time, BigInteger
 
 Base = declarative_base()
 
 class Schedule(Base):
     __tablename__ = "schedules"
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     user_id = Column(Integer, nullable=False)
     text = Column(String(1024), nullable=False)
     time = Column(Time, nullable=False)
